Add a sleep on start, add a scan function and tweak decide -> shift.
The sleep is to give X a chance to get the enter release event which
will solve some issues (enter repeat and possibly disabling a laptop
trackpad at least).  Kinda of a hack but not sure how else to solve
this, since the press and release come from different keyboards X does
not seem to connect them.
Add a --scan option which will test all the files in /dev/input to try
to find keyboards and print info about any found.
Made some changes to the decide -> shift state change.  It was possible
to have a press sent without a release (not a common case but I hit it a
couple of times) and this will cause some laptop trackpads to be
disabled (X seems to be disabling the trackpad because it thinks a key is being
pressed).